- Added interface for messaging service (message broker) - cleaned up code and reorganized folder structure
