+++ CleanArchitecture.SharedLibrary.sln
@@ -8,6 +8,7 @@ EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{BBD4C3BE-0893-457A-A6B5-C4505401A28B}"
 	ProjectSection(SolutionItems) = preProject
 		.github\workflows\dotnet.yml = .github\workflows\dotnet.yml
+		nuget.config = nuget.config
 	EndProjectSection
 EndProject
 Global
