+++ CleanArchitecture.SharedLibrary/Account/Services/UserManagement/IIdentityService.cs
@@ -1,0 +1,10 @@
+﻿using System.Threading.Tasks;
+
+namespace CleanArchitecture.SharedLibrary.Account.Services.UserManagement
+{
+    public interface IIdentityService
+    {
+        Task<string> Login(string username, string password);
+
+    }
+}
