+++ CleanArchitecture.SharedLibrary/CleanArchitecture.SharedLibrary.csproj
@@ -8,7 +8,7 @@
 	<PropertyGroup>
 		<LangVersion>10</LangVersion>
 		<Authors>Hubert Graham</Authors>
-		<Version>3.0.0</Version>
+		<Version>3.0.1</Version>
 		<Company>Kode4Hue</Company>
 		<Description>This is shared/reusable library to be used for Clean Architecture project templates</Description>
 		<GeneratePackageOnBuild>False</GeneratePackageOnBuild>
