+++ CleanArchitecture.SharedLibrary/Common/Behaviours/UnhandledExceptionBehaviour.cs
@@ -1,7 +1,6 @@
 ﻿using MediatR;
 using Microsoft.Extensions.Logging;
 using System;
-using System.Runtime.CompilerServices;
 using System.Threading;
 using System.Threading.Tasks;
