+++ CleanArchitecture.SharedLibrary/Http/Exceptions/ApiException.cs
@@ -6,7 +6,7 @@ namespace CleanArchitecture.SharedLibrary.Http.Exceptions
     [Serializable]
     public class ApiException : Exception
     {
-        public ApiException(string message) :base(message)
+        public ApiException(string message) : base(message)
         {
         }
