+++ CleanArchitecture.SharedLibrary/Http/Factories/HttpClientFactory/CustomHttpClientFactory.cs
@@ -1,10 +1,8 @@
 ﻿using CleanArchitecture.SharedLibrary.Common.Constants;
 using System;
-using System.Collections.Generic;
 using System.Net;
 using System.Net.Http;
 using System.Net.Http.Headers;
-using System.Text;
 
 namespace CleanArchitecture.SharedLibrary.Http.Factories.HttpClientFactory
 {
