+++ CleanArchitecture.SharedLibrary/Http/Factories/HttpClientFactory/ICustomHttpClientFactory.cs
@@ -1,7 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Net.Http;
-using System.Text;
+﻿using System.Net.Http;
 
 namespace CleanArchitecture.SharedLibrary.Http.Factories.HttpClientFactory
 {
