+++ CleanArchitecture.SharedLibrary/Security/Attributes/SimpleAuthorizeAttribute.cs
@@ -7,7 +7,7 @@ namespace CleanArchitecture.SharedLibrary.Security.Attributes
     /// Specifies the class this attribute is applied to requires authorization.
     /// </summary>
     [AttributeUsage(AttributeTargets.Class, AllowMultiple = true, Inherited = true)]
-    public class SimpleAuthorizeAttribute: Attribute
+    public class SimpleAuthorizeAttribute : Attribute
     {
         /// <summary>
         /// Initializes a new instance of the <see cref="AuthorizeAttribute"/> class.
