+++ CleanArchitecture.SharedLibrary/Security/Exceptions/UnauthenticatedAccessException.cs
@@ -4,7 +4,7 @@ namespace CleanArchitecture.SharedLibrary.Security.Exceptions
 {
 
     [Serializable]
-    public class UnauthenticatedAccessException: Exception
+    public class UnauthenticatedAccessException : Exception
     {
         public UnauthenticatedAccessException() : base() { }
     }
